requirements.txt
flask
flask-wtf
flask-bootstrap
rednose
nose
coverage
mongoengine
flask-mongoengine
flask-login
pyyaml
werkzeug
requests
flask_httpauth